tests: refactor uuid replacement code
Refactored code that identifies and replaces UUID variables so that:
-It is not unnecessarily ran on each loop of test.In in outer()
-outer() is not as long
